Refactor: Packaged commands
commands have been packaged and code refactored to include relevant imports
