--- src/main/java/AddTodoCommand.java
@@ -1,14 +1,0 @@
-public class AddTodoCommand extends Command {
-    private final String description;
-
-    public AddTodoCommand(String description) {
-        this.description = description;
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui) {
-        Task task = new Todo(description);
-        tasks.addTask(task);
-        ui.showAdded(task.toString(), tasks.getCount());
-    }
-}
