+++ src/main/java/commands/AddDeadlineCommand.java
@@ -1,0 +1,22 @@
+package commands;
+
+import tasks.TaskList;
+import tasks.Deadline;
+import Ui;
+
+public class AddDeadlineCommand extends Command {
+    private final String description;
+    private final String by;
+
+    public AddDeadlineCommand(String description, String by) {
+        this.description = description;
+        this.by = by;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui) {
+        Deadline task = new Deadline(description, by);
+        tasks.addTask(task);
+        ui.showAdded(task.toString(), tasks.getCount());
+    }
+}
