+++ src/main/java/commands/ExitCommand.java
@@ -1,3 +1,7 @@
+package commands;
+import tasks.TaskList;
+import Ui;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui) {
